Refactor connection beam components
