+++ src/Services/Components/ParticleEmitter.cpp
@@ -4,8 +4,6 @@
 // ============================================================================
 
 #include "Services/Components/ParticleEmitter.h"
-#include "Services/Components/ConnectionBeam.h"
-#include "Services/Components/EnergyRing.h"
 #include "Rendering/RenderContext.h"
 #include "Core/Math/MathTypes.h"
 #include "Core/Math/Math.h"
@@ -339,445 +337,4 @@ vec4 ParticleEmitter::lerpVec4(const vec4& a, const vec4& b, float t) {
     );
 }
 
-// ============================================================================
-// ConnectionBeam Implementation - Dynamic connection visualization
-// ============================================================================
-
-ConnectionBeam::ConnectionBeam() : SceneNode("Connection Beam") {
-    m_startPosition = vec3_zero();
-    m_endPosition = make_vec3(0, 0, 1);
-    m_color = make_vec3(0.4f, 0.8f, 1.0f);
-    m_intensity = 1.0f;
-    m_thickness = 0.02f;
-    m_flowSpeed = 2.0f;
-    m_flowDirection = 1.0f;
-    m_duration = 0.0f;
-    m_maxDuration = 0.0f;
-    m_isExpired = false;
-    m_connectionId = 0;
-    m_flowPhase = 0.0f;
-    m_pulsePhase = 0.0f;
-    m_segments = 32;
-    
-    createBeamGeometry();
-    createFlowEffects();
-    
-    std::cout << "ConnectionBeam created." << std::endl;
 }
-
-ConnectionBeam::~ConnectionBeam() = default;
-
-void ConnectionBeam::update(float deltaTime) {
-    SceneNode::update(deltaTime);
-    
-    updateFlow(deltaTime);
-    updatePulse(deltaTime);
-    updateDuration(deltaTime);
-    updateGeometry();
-}
-
-void ConnectionBeam::render(RenderContext& context) {
-    if (m_isExpired) return;
-    
-    context.pushTransform(getWorldMatrix());
-    context.pushBlendMode(BlendMode::ADDITIVE);
-    
-    renderBeam(context);
-    renderFlowEffects(context);
-    
-    context.popBlendMode();
-    context.popTransform();
-}
-
-void ConnectionBeam::setStartPosition(const vec3& pos) {
-    m_startPosition = pos;
-    m_geometryDirty = true;
-}
-
-void ConnectionBeam::setEndPosition(const vec3& pos) {
-    m_endPosition = pos;
-    m_geometryDirty = true;
-}
-
-void ConnectionBeam::setColor(const vec3& color) {
-    m_color = color;
-}
-
-void ConnectionBeam::setIntensity(float intensity) {
-    m_intensity = clamp(intensity, 0.0f, 2.0f);
-}
-
-void ConnectionBeam::setThickness(float thickness) {
-    m_thickness = std::max(0.001f, thickness);
-    m_geometryDirty = true;
-}
-
-void ConnectionBeam::setFlowSpeed(float speed) {
-    m_flowSpeed = speed;
-}
-
-void ConnectionBeam::setFlowDirection(float direction) {
-    m_flowDirection = direction;
-}
-
-void ConnectionBeam::setDuration(float duration) {
-    m_maxDuration = duration;
-    m_duration = 0.0f;
-    m_isExpired = false;
-}
-
-void ConnectionBeam::setConnectionId(uint32_t id) {
-    m_connectionId = id;
-}
-
-void ConnectionBeam::pulse(float intensity) {
-    m_pulsePhase = 0.0f;
-    m_pulseIntensity = intensity;
-}
-
-bool ConnectionBeam::isExpired() const {
-    return m_isExpired;
-}
-
-// Private implementation methods
-
-void ConnectionBeam::createBeamGeometry() {
-    // Create beam mesh with segments for smooth curves
-    m_beamMesh = std::make_shared<BeamMesh>();
-    m_beamMesh->setSegmentCount(m_segments);
-    updateBeamMesh();
-    
-    // Create beam material
-    m_beamMaterial = std::make_shared<Material>("connection_beam");
-    m_beamMaterial->setAlbedo(m_color);
-    m_beamMaterial->setEmission(m_color);
-    m_beamMaterial->setEmissionStrength(m_intensity);
-    m_beamMaterial->setAlpha(0.8f);
-}
-
-void ConnectionBeam::createFlowEffects() {
-    // Create flow particle system
-    ParticleEmitter::Config flowConfig;
-    flowConfig.emitShape = ParticleEmitter::Shape::LINE;
-    flowConfig.emitRate = 20.0f;
-    flowConfig.particleLifetime = 1.0f;
-    flowConfig.startSize = 0.015f;
-    flowConfig.endSize = 0.005f;
-    flowConfig.startColor = make_vec4(m_color.x, m_color.y, m_color.z, 1.0f);
-    flowConfig.endColor = make_vec4(m_color.x, m_color.y, m_color.z, 0.0f);
-    flowConfig.velocity = 3.0f;
-    flowConfig.inheritVelocity = 0.8f;
-    
-    m_flowParticles = std::make_shared<ParticleEmitter>(flowConfig);
-    m_flowParticles->setName("Beam Flow");
-    addChild(m_flowParticles);
-    
-    updateFlowLine();
-}
-
-void ConnectionBeam::updateFlow(float deltaTime) {
-    m_flowPhase += deltaTime * m_flowSpeed * m_flowDirection;
-    
-    // Update flow particle emission
-    if (m_flowParticles) {
-        updateFlowLine();
-        
-        // Modulate emission based on intensity
-        float emissionRate = 15.0f * m_intensity;
-        m_flowParticles->setEmitRate(emissionRate);
-        
-        // Update particle color
-        vec4 flowColor = make_vec4(m_color.x, m_color.y, m_color.z, 0.8f * m_intensity);
-        m_flowParticles->setParticleColor(flowColor);
-    }
-}
-
-void ConnectionBeam::updatePulse(float deltaTime) {
-    if (m_pulseIntensity > 0.0f) {
-        m_pulsePhase += deltaTime * 6.0f;
-        m_pulseIntensity *= pow(0.95f, deltaTime * 60.0f); // Exponential decay
-    }
-}
-
-void ConnectionBeam::updateDuration(float deltaTime) {
-    if (m_maxDuration > 0.0f) {
-        m_duration += deltaTime;
-        if (m_duration >= m_maxDuration) {
-            m_isExpired = true;
-        }
-    }
-}
-
-void ConnectionBeam::updateGeometry() {
-    if (m_geometryDirty) {
-        updateBeamMesh();
-        updateFlowLine();
-        m_geometryDirty = false;
-    }
-}
-
-void ConnectionBeam::updateBeamMesh() {
-    if (!m_beamMesh) return;
-    
-    // Calculate beam parameters
-    vec3 direction = m_endPosition - m_startPosition;
-    float length = magnitude(direction);
-    vec3 normalizedDir = normalize(direction);
-    
-    // Generate beam vertices with slight curve for organic feel
-    std::vector<vec3> centerLine;
-    for (int i = 0; i <= m_segments; ++i) {
-        float t = i / float(m_segments);
-        vec3 point = lerp(m_startPosition, m_endPosition, t);
-        
-        // Add subtle curve
-        float curvature = sin(t * M_PI) * length * 0.02f;
-        vec3 perpendicular = cross(normalizedDir, make_vec3(0, 1, 0));
-        if (magnitude(perpendicular) < 0.1f) {
-            perpendicular = cross(normalizedDir, make_vec3(1, 0, 0));
-        }
-        perpendicular = normalize(perpendicular);
-        
-        point = point + perpendicular * curvature;
-        centerLine.push_back(point);
-    }
-    
-    m_beamMesh->updateCenterLine(centerLine, m_thickness);
-}
-
-void ConnectionBeam::updateFlowLine() {
-    if (m_flowParticles) {
-        m_flowParticles->setEmitLine(m_startPosition, m_endPosition);
-        
-        // Set flow direction
-        vec3 flowDir = normalize(m_endPosition - m_startPosition);
-        if (m_flowDirection < 0.0f) {
-            flowDir = -flowDir;
-        }
-        m_flowParticles->setEmitDirection(flowDir);
-    }
-}
-
-void ConnectionBeam::renderBeam(RenderContext& context) {
-    if (!m_beamMesh) return;
-    
-    // Calculate current intensity with pulse
-    float currentIntensity = m_intensity;
-    if (m_pulseIntensity > 0.0f) {
-        currentIntensity += m_pulseIntensity * sin(m_pulsePhase);
-    }
-    
-    // Apply fade-out if expiring
-    if (m_maxDuration > 0.0f && m_duration > 0.0f) {
-        float fadeRatio = 1.0f - (m_duration / m_maxDuration);
-        currentIntensity *= fadeRatio;
-    }
-    
-    // Set beam color and intensity
-    vec4 beamColor = make_vec4(m_color.x, m_color.y, m_color.z, currentIntensity);
-    context.setColor(beamColor);
-    context.setEmission(m_color * currentIntensity);
-    
-    // Render beam mesh
-    m_beamMesh->render(context);
-}
-
-void ConnectionBeam::renderFlowEffects(RenderContext& context) {
-    // Flow effects are rendered by the particle system automatically
-    // Additional flow visualization could be added here
-    
-    // Render flow indicators along the beam
-    if (m_intensity > 0.5f) {
-        renderFlowIndicators(context);
-    }
-}
-
-void ConnectionBeam::renderFlowIndicators(RenderContext& context) {
-    // Render small indicators moving along the beam
-    int indicatorCount = 3;
-    vec3 direction = m_endPosition - m_startPosition;
-    float length = magnitude(direction);
-    
-    for (int i = 0; i < indicatorCount; ++i) {
-        float offset = (i / float(indicatorCount)) + m_flowPhase * 0.1f;
-        offset = fmod(offset, 1.0f);
-        
-        if (m_flowDirection < 0.0f) {
-            offset = 1.0f - offset;
-        }
-        
-        vec3 indicatorPos = m_startPosition + direction * offset;
-        
-        context.pushTransform(mat4_identity());
-        context.translate(indicatorPos);
-        
-        float indicatorIntensity = m_intensity * 0.8f;
-        vec4 indicatorColor = make_vec4(m_color.x, m_color.y, m_color.z, indicatorIntensity);
-        context.setColor(indicatorColor);
-        
-        context.drawGlowSphere(0.03f);
-        
-        context.popTransform();
-    }
-}
-
-// ============================================================================
-// EnergyRing Implementation - Rotating energy rings for the nexus
-// ============================================================================
-
-EnergyRing::EnergyRing() : SceneNode("Energy Ring") {
-    m_innerRadius = 2.0f;
-    m_outerRadius = 2.2f;
-    m_height = 0.1f;
-    m_color = make_vec3(0.4f, 0.8f, 1.0f);
-    m_glowIntensity = 1.0f;
-    m_rotationSpeed = 1.0f;
-    m_flowDirection = 1.0f;
-    m_segments = 64;
-    m_pulseIntensity = 0.0f;
-    m_pulsePhase = 0.0f;
-    m_rotationPhase = 0.0f;
-    m_rotationAxis = make_vec3(0, 1, 0);
-    m_distortionAmount = 0.0f;
-    
-    createRingGeometry();
-    createRingEffects();
-    
-    std::cout << "EnergyRing created with radius: " << m_innerRadius << std::endl;
-}
-
-EnergyRing::~EnergyRing() = default;
-
-void EnergyRing::update(float deltaTime) {
-    SceneNode::update(deltaTime);
-    
-    updateRotation(deltaTime);
-    updatePulse(deltaTime);
-    updateDistortion(deltaTime);
-    updateEffects(deltaTime);
-}
-
-void EnergyRing::render(RenderContext& context) {
-    context.pushTransform(getWorldMatrix());
-    context.pushBlendMode(BlendMode::ADDITIVE);
-    
-    renderRing(context);
-    renderGlow(context);
-    
-    context.popBlendMode();
-    context.popTransform();
-}
-
-void EnergyRing::setInnerRadius(float radius) {
-    m_innerRadius = std::max(0.1f, radius);
-    m_geometryDirty = true;
-}
-
-void EnergyRing::setOuterRadius(float radius) {
-    m_outerRadius = std::max(m_innerRadius + 0.05f, radius);
-    m_geometryDirty = true;
-}
-
-void EnergyRing::setHeight(float height) {
-    m_height = std::max(0.01f, height);
-    m_geometryDirty = true;
-}
-
-void EnergyRing::setColor(const vec3& color) {
-    m_color = color;
-    updateMaterialColor();
-}
-
-void EnergyRing::setGlowIntensity(float intensity) {
-    m_glowIntensity = clamp(intensity, 0.0f, 2.0f);
-}
-
-void EnergyRing::setRotationSpeed(float speed) {
-    m_rotationSpeed = speed;
-}
-
-void EnergyRing::setFlowDirection(float direction) {
-    m_flowDirection = direction;
-}
-
-void EnergyRing::setRotationAxis(const vec3& axis) {
-    m_rotationAxis = normalize(axis);
-}
-
-void EnergyRing::triggerRipple(float intensity) {
-    // Create ripple effect
-    auto ripple = std::make_shared<RingRipple>();
-    ripple->setRadius(m_innerRadius);
-    ripple->setIntensity(intensity);
-    ripple->setColor(m_color);
-    ripple->setDuration(1.5f);
-    
-    addChild(ripple);
-    
-    // Trigger pulse
-    setPulseIntensity(intensity);
-    
-    std::cout << "Ring ripple triggered with intensity: " << intensity << std::endl;
-}
-
-void EnergyRing::setPulseIntensity(float intensity) {
-    m_pulseIntensity = clamp(intensity, 0.0f, 2.0f);
-    m_pulsePhase = 0.0f;
-}
-
-void EnergyRing::setDistortionAmount(float amount) {
-    m_distortionAmount = clamp(amount, 0.0f, 1.0f);
-}
-
-// Private implementation methods
-
-void EnergyRing::createRingGeometry() {
-    // Create ring mesh
-    m_ringMesh = std::make_shared<RingMesh>();
-    m_ringMesh->setInnerRadius(m_innerRadius);
-    m_ringMesh->setOuterRadius(m_outerRadius);
-    m_ringMesh->setHeight(m_height);
-    m_ringMesh->setSegments(m_segments);
-    m_ringMesh->build();
-    
-    // Create ring material
-    m_ringMaterial = std::make_shared<Material>("energy_ring");
-    m_ringMaterial->setAlbedo(m_color);
-    m_ringMaterial->setEmission(m_color);
-    m_ringMaterial->setEmissionStrength(m_glowIntensity);
-    m_ringMaterial->setAlpha(0.7f);
-    m_ringMaterial->setMetallic(0.1f);
-    m_ringMaterial->setRoughness(0.2f);
-}
-
-void EnergyRing::createRingEffects() {
-    // Create energy flow particles around the ring
-    ParticleEmitter::Config flowConfig;
-    flowConfig.emitShape = ParticleEmitter::Shape::CIRCLE;
-    flowConfig.emitRadius = (m_innerRadius + m_outerRadius) * 0.5f;
-    flowConfig.emitRate = 25.0f;
-    flowConfig.particleLifetime = 2.0f;
-    flowConfig.startSize = 0.02f;
-    flowConfig.endSize = 0.005f;
-    flowConfig.startColor = make_vec4(m_color.x, m_color.y, m_color.z, 0.8f);
-    flowConfig.endColor = make_vec4(m_color.x, m_color.y, m_color.z, 0.0f);
-    flowConfig.velocity = 1.5f;
-    
-    m_flowParticles = std::make_shared<ParticleEmitter>(flowConfig);
-    m_flowParticles->setName("Ring Flow");
-    addChild(m_flowParticles);
-    
-    // Create spark effects
-    createSparkEffects();
-}
-
-void EnergyRing::createSparkEffects() {
-    // Occasional sparks along the ring
-    ParticleEmitter::Config sparkConfig;
-    sparkConfig.emitShape = ParticleEmitter::Shape::CIRCLE;
-    sparkConfig.emitRadius = m_outerRadius;
-    sparkConfig.emitRate = 5.0f;
-    sparkConfig.particleLifetime = 0.8f;
-    sparkConfig.startSize = 0.03f;
-    sparkConfig.endSize = 0.01f;
-    sparkConfig.startColor = make_vec4(1.0f, 1.0f, 1.0
